feat: Update dependencies and add Vuestic UI library
This commit updates the dependencies in the `package.json` and `pnpm-lock.yaml` files to include the latest stable versions of `@vueuse/core`, `@vueuse/firebase`, `vue`, `vuefire`, and `vuestic-ui`. It also adds the `vuestic-ui` library as a new dependency. The `resolutions` field in `package.json` is also updated to include an alias for the `@` symbol, pointing to the `./src` directory. This will make it easier to import files from the `src` directory using the `@` alias.